+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/DrawerAttributes/ShowMessageAttribute.cs
@@ -1,14 +1,13 @@
 ﻿using System;
 using UnityEditor;
-using UnityEngine;
 
-namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.DrawerAttributes
 {
     /// <summary>
     /// Allows to display a box above the field to give a warning or info
     /// </summary>
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = true, Inherited = true)]
-    public class ShowMessageAttribute : PropertyAttribute, ILaborerAttribute
+    public class ShowMessageAttribute : DrawerAttribute
     {
 
         public string Message { get; private set; }
@@ -16,8 +15,8 @@ namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
 
         public ShowMessageAttribute(string message, MessageType messageType = MessageType.None)
         {
-            this.Message = message;
-            this.MessageType = messageType;
+            Message = message;
+            MessageType = messageType;
         }
     }