+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/InspectedAttributes/LabelAttribute.cs
@@ -1,0 +1,12 @@
+﻿namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.InspectedAttributes
+{
+    public abstract class LabelAttribute: InspectedAttribute
+    {
+        public readonly string _label;
+
+        protected LabelAttribute(string label)
+        {
+            _label = label;
+        }
+    }
+}