+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/InspectedAttributes/ReadOnlyAttribute.cs
@@ -1,13 +1,12 @@
 using System;
-using UnityEngine;
 
-namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.InspectedAttributes
 {
     /// <summary>
     /// Allows to display an attribute as ReadOnly
     /// </summary>
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = false, Inherited = true)]
-    public class ReadOnlyAttribute : PropertyAttribute, ILaborerAttribute
+    public abstract class ReadOnlyAttribute : InspectedAttribute
     {
     }