+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/InspectedAttributes/ShowIfAttributeBase.cs
@@ -1,33 +1,34 @@
 ﻿using System;
 using GraphicsLabor.Scripts.Attributes.Utility;
 
-namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.InspectedAttributes
 {
-    public class ShowIfAttributeBase : Attribute, ILaborerAttribute
+    public class ShowIfAttributeBase : InspectedAttribute
     {
         public string[] Conditions { get; private set; }
         public ConditionOperator ConditionOperator { get; private set; }
         public bool Inverted { get; protected set; }
         public Enum EnumValue { get; private set; }
 
-        public ShowIfAttributeBase(string condition)
+        // TODO: Apply ShowIf to Properties
+        protected ShowIfAttributeBase(string condition)
         {
             ConditionOperator = ConditionOperator.And;
             Conditions = new[] { condition };
         }
 
         // Allows for showing a field if certains conditions are met (bools are true)
-        public ShowIfAttributeBase(ConditionOperator conditionOperator, params string[] conditions) // params string[] equivalent to *args of python but strongly typed
+        protected ShowIfAttributeBase(ConditionOperator conditionOperator, params string[] conditions) // params string[] equivalent to *args of python but strongly typed
         {
-            this.ConditionOperator = conditionOperator;
-            this.Conditions = conditions;
+            ConditionOperator = conditionOperator;
+            Conditions = conditions;
         }
 
         // Allows for showing a field if an enum field has a certain value
-        public ShowIfAttributeBase(string enumName, Enum enumValue)
+        protected ShowIfAttributeBase(string enumName, Enum enumValue)
             : this(enumName) // Like in C++ calls ShowAttributeBase before
         {
-            this.EnumValue = enumValue ?? throw new ArgumentNullException(nameof(enumValue), "This parameter must be an enum value.");
+            EnumValue = enumValue ?? throw new ArgumentNullException(nameof(enumValue), "This parameter must be an enum value.");
         }
     }
 }