+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/InspectedAttributes/ShowPropertyAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes.InspectedAttributes
+{
+    [AttributeUsage(AttributeTargets.Property, AllowMultiple = false, Inherited = true)]
+    public class ShowPropertyAttribute : InspectedAttribute
+    {
+        
+    }
+}