+++ Assets/GraphicsLabor/Scripts/Attributes/Utility/Enums.cs
@@ -2,7 +2,6 @@
 {
     public enum ConditionOperator
     {
-        // Maybe do a NOR if possible
         And,
         Or
     }