+++ CHANGELOG.md
@@ -6,6 +6,36 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [3.0.0] - 2023-11-02
+
+### Added
+
+- Editable Attribute for SO
+- TabProperty Attribute for SO fields and Properties with the ShowIf Attribute that have the Editable Attribute
+- Manager Attribute for SO
+- ObjectExtension, extension for UnityEngine.Object to test for Inheritance of Types and GetInheritedTypes
+- LaborerGUIUtility, a class to hold needed constants for Drawing GUIs
+- AssetHandler, handles double clicking on supported assets
+- EditorWindowBase, Base class for custom EditorWindows, implements functionality such as opening various windows with the same type
+- SciptableObjectEditorWindow, window that enhances scriptable objects' edit
+- ShowIf Attributes also works with TabProperty on Fields
+- GetAttributes overload for PropertyInfo
+
+
+### Changed
+
+- Code Refactoring, mainly changes in method scopes and cleaned imports
+- ExpandableDrawer now uses LaborerGUIUtility for spacing
+- HorizontalSeparatorDrawer now uses LaborerGUIUtility for spacing
+- GetSelfAndBaseTypes is replaced by Object.GetTypes()
+- Updated CHANGELOG
+
+
+### Know Bugs
+
+- Opening a custom Editor window after recompilation always creates a new window
+
+
 
 ## [2.0.0] - 2023-10-29